+++ filter.hpp
@@ -10,27 +10,27 @@
 #include <stdlib.h>
 
 #define DEFAULT_FC (1000)
-#define DEFAULT_Q (1.0f)
+#define DEFAULT_Q (1.0)
 #define DEFAULT_ORDER (2)
 
 using namespace std;
 
-// idee : créer un template pour travailler en float ou double
+// idee : créer un template pour travailler en double ou double
 
 enum filtType {LPF, HPF, BPF};
 
 class Filter{
 private:
 
-  float coeff[6];   //6 coefficients for biquad filter design
+  double coeff[6];   //6 coefficients for biquad filter design
   filtType filterType; //lowpass, bandpass, highpass
   uint32_t Fc; //cut-off frequency or center frequency for BPF case
   uint32_t Fs; //sampling frequency
   uint32_t bw;
   uint32_t order; //filter order (2 or 4)
-  float Q; //quality factor
-  float alpha;
-  float wc; //normalized cut-off frequency
+  double Q; //quality factor
+  double alpha;
+  double wc; //normalized cut-off frequency
 
   void DesignFilter( filtType _filterType);
   // filling filter coefficients array regarding filter type and parameters given
@@ -42,7 +42,7 @@ private:
   void setType_Init(filtType newFilterType);
   void setFc_Init(uint32_t newFc);
   void setBandwidth_Init(uint32_t newBw);
-  void setQ_Init(float newQ);
+  void setQ_Init(double newQ);
   void setOrder_Init(uint32_t newOrder);
 
 
@@ -51,8 +51,8 @@ public:
   //constructors
   Filter();
 
-  Filter(filtType _filterType, uint32_t _Fc, float _Q, uint32_t _order);
-  Filter(filtType _filterType, uint32_t _Fc, float _Q, uint32_t _bw, uint32_t _order);
+  Filter(filtType _filterType, uint32_t _Fc, double _Q, uint32_t _order);
+  Filter(filtType _filterType, uint32_t _Fc, double _Q, uint32_t _bw, uint32_t _order);
 
 
   //destructors
@@ -63,16 +63,16 @@ public:
   void setType(filtType newFilterType);
   void setFc(uint32_t newFc);
   void setBandwidth(uint32_t newBw);
-  void setQ(float newQ);
+  void setQ(double newQ);
   void setOrder(uint32_t newOrder);
 
   //getter
   filtType getType();
   uint32_t getFc();
   uint32_t getBandwidth();
-  float getQ();
+  double getQ();
   uint32_t getOrder();
-  float getCoeff(uint32_t i);
+  double getCoeff(uint32_t i);
 
   //other functions
   void filterArrayCompute(int16_t* iarray, int16_t* oarray, uint32_t iLen);
