+++ oscillators.cpp
@@ -48,9 +48,9 @@ void Osc::setWF(int16_t * _wavetable) {
 }
 
 int16_t Osc::linearInterpolation() {
-  float temp = phase * PHASE_FACTOR; // Scales the phase to LUT resolution
+  double temp = phase * PHASE_FACTOR; // Scales the phase to LUT resolution
   int16_t intPart = temp; // int part of the scaled phasor
-  float fracPart = temp-intPart; // fractional part
+  double fracPart = temp-intPart; // fractional part
   int16_t sample0 = wavetable[intPart]; // first sample
   if(++intPart >= RES) // modulo
     intPart = 0;
